+++ addOns/pscan/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.4.0] - 2025-09-02
 ### Added
 - Allow to configure the option max body size through the API (Issue 8974).
 - Support for stopping the passiveScan-wait automation job.
@@ -66,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `passiveScan-wait`
 - Dynamically un/load add-on passive scan rules (Issue 7959).
 
+[0.4.0]: https://github.com/zaproxy/zap-extensions/releases/pscan-v0.4.0
 [0.3.0]: https://github.com/zaproxy/zap-extensions/releases/pscan-v0.3.0
 [0.2.1]: https://github.com/zaproxy/zap-extensions/releases/pscan-v0.2.1
 [0.2.0]: https://github.com/zaproxy/zap-extensions/releases/pscan-v0.2.0